requirements.txt
pymupdf
google-generativeai
chromadb
python-dotenv
pytest
pytest-cov
langchain-core
langchain-google-genai
langchain-community
langchain
langchain-chroma
langchain-text-splitters
pydantic
tiktoken